fix(membership): update service.remove
to verify that channel was found before checking channel.type

+++ backend/src/memberships/memberships.service.ts
@@ -62,7 +62,7 @@ export class MembershipsService {
     });
     if (
       membership.role === MembershipRoleType.OWNER ||
-      channel.type === ChannelType.DIRECT
+      (channel && channel?.type === ChannelType.DIRECT)
     ) {
       this.channelsRepository.delete(channel.id);
     }
